Implement permissions for checking conflicts
This commit implements the conflicts permissions to view and edit
conflicts from the admin area, disabling all fields and removing the
"Delete" checkboxes if in read-only.

As a consequence of adding the permissions check to the admin mixin, all
views will be disabled for non-superusers, until a permission has been
added to the view, or the permission was set to `True`.