Create path to access conflicts for auth'd users
This commit creates a path for authenticated users to click through
links on Tabbycat until they reach a page which requires authorization
(currently the conflicts pages), without encountering a 401/403.

This means that the admin overview and Simple Importer index pages need
to be accessible for all logged-in users, although all data seen in the
overview page would be hidden behind permissions.

For this, many new permissions have been added from those which are used
in the overview, plus some extra ones that sprang to mind. These haven't
been added to the migration yet, but choices are not set at the DB level
anyways.