API: Allow answers to be in any type
This commit creates a new field which takes in a value of Any type and
returns it, to allow for the polymorphism of adjudicator feedback
questions' answers. The correct type for the question is then checked in
validation, or rejected otherwise.

This fixes the bad handling of multiple choice questions as arrays are
now permitted and well deserialized.

Extra validations are also added to make sure the answers are suitable
(in the available choices) and that all required questions are answered.