Add validation between settings
This commit replaces the after-the-fact warning messages about having
incompatible settings with form validations that prevent saving the
offending settings. With this new approach, the error messages can be
shorter, as they are shown with the setting(s) to be changed.

An improvement could be if the validations could be separated by
section, to avoid listing them all in the same place, disorganized.

There can be more validations added later in the same place and format.
The current validations are:
- Adj/Speaker [reply] score minimums must be less than their maximums
- Intermediate 1/2 brackets require preallocated sides
- BP can't use voting ballots
- Institutions shouldn't be shown with team names