Add ability to modify preset applications
This commit reconfigures the preset application confirmation page to
allow for users to change the values for the settings that will be set.
This would make the page more useful as a check that a proper
configuration was made, and departures from the preset being easier and
more explicit. For space and style, the previous values have been
removed, but the page brought more in line with other forms.

The new page was accomplished by making the page into a form view, with
the preset options rendered as fields. For this to work effectively, the
functions used with presets have been refactored, preferring class
methods within the Base class. These new methods allow for the list of
settings to be retrieved, along with making the preference form and
saving the preset directly.

With the refactor, places that used the old functions have been migrated
and in some places (applypreset command), simplified.

Trying to get into the habit of unit testing!