Create form for inviting a user
This commit creates a tournament-scoped form view which extends the
password reset feature, as to invite a user to create an account with
the given permissions. This works by creating them an account with their
email address, and then the "password reset" link will prompt them to
also choose a username.

The new email template hasn't been created yet, nor their form view to
create their account. If a user already exists, an email should still be
sent, but with the reset link redirecting to the inviting tournament.

This form is designed to be extensible to tournament-based granular
permissions, through having a "role" field and being tournament-scoped.