Create mechanism for grouping permissions
This commit creates the models and admin integration to be able to
create groupings of permissions and assign them to users. This would
facilitate giving standard permissions to different users by the role(s)
they have per tournament.

Groups are lists of permissions, applied to a specific tournament. This
allows groups to be modified without affecting permissions elsewhere.
Users are then assigned these groups as Memberships. Individual
permissions can also be added through UserPermission objects.

There will be default lists of permissions that will form groups on the
creation of tournaments:
* Equity: Manage conflicts/constraints
* Adjudicator Core: Create allocations and motions

In addition to a Tab Director role with full access, and "Tab Assistant"
with equivalent permissions to what is available on the assistant
interface.

The DB page for users has also been updated to be localisable, and be
able to assign membership and permissions.