Create new type for automated submissions
This commit creates a new submission type for when Tabbycat creates
ballots or feedback automatically. This is in view of creating ballots
automatically for byes. This is required for creating TeamScore objects,
as they would get a win.

It also updates the choices list to be an enumeration (as TextChoices),
which is Django's new style.