Associate permissions with views
This commit sets the newly-added permissions with the views that they'll
restrict. It also makes some small modifications to the enumeration.

We also set the view permission to be the edit permission if it is set,
for views which are useless if cannot post (like adding feedback).

It also adds simple typings and fixes a bug where the user object's set
of permissions could only be size of 1.