Use speaker ranks as part of scoresheet
This commit consolidates the handling of speaker ranks as part of the
result scoresheet rather than the result itself, as both methods were
being used and led to confusion, with the ranks not displaying nor
saving.

It also fixes a typo with how the feature showing is determined and
simplifies some duplicated methods.