API: Fix couple problems when creating results
This commit fixes adding vetos to a ballot through the API and now
allows for voting ballots to be created, single_adj or full. We can then
use the proper score-setting method signature with[out] adjudicator.

Tests are also now provided for this POST endpoint, silencing the logger
as the tests do not expect any.