Add date to round start time
This commit migrates the Round "start time" field to include the date
(and timezone) in order to better support prepared rounds. The generated
migration will take the site's timezone and the first ballot that was
submitted to create a DateTime.

The "set start time" modal now has an initial value of 15 minutes after
the current date-time, and the date is shown. The API will accept both
time and datetime objects.