Tests: Correct expected values for draw strength
This commit places back the tests for draw strength when faced with a
unconfirmed ballot. Contrary to other metrics, the values should change
as they include the new draw. However, they should not contain the
scores from the unconfirmed ballot. This test shows that this behaviour
is current, and is expected.