Convert sides in debates to a sequence
This commit starts preparations for having debates with an arbitrary
number of teams, for use with public speaking competitions. Rather than
have a pre-set list of options for sides (aff/neg, og/oo/cg/co), "teams"
(loose definition as public speaking teams are individual speakers) will
have a sequence number for which order their team will speak first.

This commit generalizes the "teams in debate" preference to allow any
number of teams rather than 2 or 4. It then converts all the side
choices to an IntEnum for sequence numbers.

For API backwards compatibility, even though now the list of teams is
ordered by side sequence, a "side" field is kept, calculating what the
side key would have been, if possible.

This commit does not implement generating debates with different sizes,
only makes it possible to store, and be able to handle this new format
for the existing side numbers.