Keep order of debates in private URL ballot column
This commit keeps the order of debates for ballot links as given,
despite the debate query being replaced by another one. The objects from
the new query get put into a dict to be accessed in the iteration of
debates, in order to keep them sorted.

The alternative was to sort the query by round sequence, but that would
be less generalizable.

Fixes #2369