API: Use mixin to log actions
This commit prepares API endpoints to use the actionlog, through
adapting the logging mixin to the tournament API mixin. As the latter is
already widely used, minimal code will be required to set the action
type.

It also simplifies the lookup kwargs to be just the "nearest" object
so that `perform_create()` does not have to be overridden. This should
not break anything as their use with tournament, round, and debate have
foreign keys to each other.