Add "agent" field to action log
This commit adds an "agent" field to action log entries, to mirror the
"agent" of Activity Streams. It will be used to identify which actions
have been taken through the API vs on web.

Future development could further add agents, such as "through private
URL" or "admin/assistant side". This potential is why it's added as a
choice field rather than boolean.