API: Add barcodes to participants
This commit adds participants' checkin barcodes to the participants'
list and detail endpoints to allow for them to be easily exported
without needing to make one request per participant.

This field is private, and currently read-only, but can be opened up.

It also fixes the type for the barcode in the main checkin serializer,
as barcodes are strings rather than numbers (e.g. allowing "000001").