API: Keep track of base score history for adjs
This commit adds adjudicator base score tracking to the API. Base score
changes are tracked within the adjfeedback app with the related current
round, to keep the visualizations constant. These changes are tracked
with the AdjudicatorBaseScoreHistory model.

Now, when updating (PUT/PATCH) an adjudicator and the base score has
been modified, a new history object is created.