Specify hidden value for tournament in break cats
This commit adds the initial value for the tournament (hidden field) in
the break category formset. It being absent made the formset fail to
create new break categories with a general error message.

The reliance on the hidden field should be removed in the future, with
uniqueness checks done with the tournament attribute.