API: Add motion seq field back for rounds
The sequence field had been removed in the Round API endpoint in favour
of relying on the ordering in the array. However, as the field was still
in use, it caused an unintended breakage. The sequence will now remain,
but be read-only as we look to the ordering when creating them.