Enable creating results for public speaking
This commit extends the BP results framework to allow for any number of
teams, renaming BP-specific scoresheets to "Poly;" relaxing side numbers
to the amount specified by the tournament.

The ballot form is also updated to use a grid of scoresheets, and using
the results table to not use chevrons.