Implement score criteria within ballots
This commit implements the use of score criteria within the old ballot
form, adding fields in the result forms for both consensus and voting
formats. The handling is then done within DebateResult and Scoresheet.

If criteria are detected for the tournament, the regular speaker score
fields become a weighted sum (weight added as an attribute of the
criterion fields) rather than directly editable, calculated on the form
with JS and calculated in the scoresheet.