Create auth scheme for private URL keys
This commit creates a new authentication class that takes an
"Authorization" header with a private URL key (prefixed by "Key"), to
identify participants with their use of API endpoints where appropriate.

It will be necessary for participants to use the API with private keys
when the new frontend will be in use. Extra permissions logic has been
added to ensure participants only have access to their data, to act as
themselves, and only if the relevant preference has been set.

A couple endpoints have been adapted:
* Ballots:
  One's own ballots can always be seen, but preference on whether to
  create them. Teams can see theirs on the "private_ballots_released"
  setting.
* Feedback: Participants can always see the feedback they have submitted
  but not whether it is ignored. If "participant_ballots" set, either no
  auth or private URL key is needed.
* Check-in: Participants can check themselves in and out if self-check-
  in active.

Added a fake doc comment to avoid the LogActionMixin text from appearing
on all endpoint docs.