Refactor feedback questions to use content types
In order to be able to use the existing question/answer adjudicator
feedback framework for participant registration, a few things had to be
done:
* Create a "Question" model:
  The new Question model is split off from the existing Adjudicator
  Feedback Question model, using a OneToOneField to link. The new
  question model takes all the general fields, and adds a foreign key
  for the content type the question would apply to.
* Add content type fields for answers
  Answers to questions can now refer to any model, so a content type FK
  was added, renaming the feedback ID to "object ID" for use with any
  object. This is not normalized, but is simpler to use and extend if
  ever questions can be used for many models. These have also been
  renamed to be general.

A custom migration is created, taking care to retain all IDs. Content
type IDs have to be queried.

The admin view, API, and feedback views are updated to use this new
system, including prefetching to be efficient.