Create "draw strength by rank" metric annotator
This commit adds a new team metric that takes the opponent teams' ranks
as a sum, for use in WSDC draw pull-up rules.

As the metric relies on the rankings, this metric must be run after
having sorted the precedence metrics, and cannot be used itself for
ranking. To solve these, now, metrics can be excluded from the options
ranking metrics lists, and the check for adding metrics after sorting is
removed.