API: Create Full tournament export endpoint
As a replacement for DebateXML, we could attempt to use a nested
structure with our existing API serializers as a base for the export and
archival of Tabbycat data.

This approach will improve maintainability as more functionality can be
achieved with less code, and that code is re-usable.

The challenge is how to handle imports through this endpoint, as it
would have to be possible to import data with lateral foreign keys (say
participants in a pairing), but these keys would be different once
created.