Allow ballots to be marked as forfeited
This commit re-adds a "forfeit" feature, allowing adjudicators to mark a
team as having forfeited a round. The effect is that the opposing team
gets a win, but no speaker scores are attributed (and so requires the
use of average speaks rather than total).

For this, a new field in BallotSubmission is added to short-circuit the
DebateResult and form classes, redirecting the winner attribute to the
"forfeit" checkboxes.